Revamp asset build process, split asset manifest generation into own step
This is a framework for outputting arbitrary number of cooked assets from a single
source asset. For example multiple meshes, textures, animations, etc. from a single gltf file.

+++ src/assets/root.zig
@@ -1,4 +1,4 @@
-pub const AssetId = u32;
+pub const AssetId = u64;
 
 pub const Handle = struct {
     pub const Shader = extern struct { id: AssetId = 0 };
